Add pathogen installation instructions
I added pathogen installation instructions to the README.md. In doing so, I had
to change some of the instructions that were already there. I also changed the
master headings to try to keep everything clean and easy to read.

I don't have Vundle installed, so I was unable to test to make sure those
instructions were correct. Otherwise, I tested both the copy and paste
instructions and the Pathogen instructions. A second test, at least of the copy
and paste instructions, would be great.

dlpolanco